pointerlock/constructor.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8">
    <title>Pointer Lock event constructor</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>  
    <link rel="help" href="http://www.w3.org/TR/pointerlock/#pointerlockchange-and-pointerlockerror-events"/>
    <link rel="help" href="http://www.w3.org/TR/dom/#interface-event"/>
  </head>
  <body>
    <p>This test the events for Pointer Lock can be created as defined by DOM events.</p>
    <div id='log'></div>
    <script>
test(function() {
  var ev = new MouseEvent("pointerlockchange");
  assert_equals(ev.type, "pointerlockchange");
  assert_equals(ev.target, null);
  assert_equals(ev.currentTarget, null);
  assert_equals(ev.bubbles, false);
  assert_equals(ev.eventPhase, Event.NONE);
  assert_equals(ev.cancelable, false);
  assert_true("preventDefault" in ev);
  assert_equals(ev.defaultPrevented, false);
  assert_true(ev.timeStamp > 0);
  assert_true("initEvent" in ev);
  assert_true("movementX" in ev, "movementX exists");
  assert_true("movementY" in ev, "movementY exists");
  assert_true(ev.movementX == 0, "movementX is 0");
  assert_true(ev.movementY == 0, "movementY is 0");
}, "Default event values for mouse event interface and its pointer lock extensions.");
test(function() {
  var ev = new MouseEvent("pointerlockerror", { type: "trololol", bubbles: true, cancelable: false, defaultPrevented: true});
  assert_equals(ev.type, "pointerlockerror");
  assert_equals(ev.bubbles, false);
  assert_equals(ev.cancelable, false);
  assert_equals(ev.defaultPrevented, false);
}, "Basic test.");
    </script>

</body>

</html>
